requirements.txt
pydantic
firecrawl-py
crewai
crewai-tools
python-dotenv
langchain
langdetect
deep-translator